requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
google-auth
requests
python-dotenv